requirements/requirements.txt
asgiref==3.3.1
cffi==1.14.5
cryptography==3.4.7
Django==3.1.5
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-js-asset==1.2.2
django-map-widgets==0.3.0
django-menu-generator==1.1.0
django-model-utils==4.1.1
djangorestframework==3.12.2
gunicorn==20.0.4
importlib-metadata==3.4.0
Markdown==3.3.3
Pillow==8.1.0
psycopg2-binary==2.8.6
pycparser==2.20
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
zipp==3.4.0
